I finally fixed the servo_movement package (after at least 20 days of tinkering).
I feel a bit stupid because the solution was very simple, I had always used 'ros2 launch' to launch my packages, while in this type of case where we take I/O, we must use 'ros2 run'. It seems obvious when you think about it. This obviousness was gone when I started to almost get it to work using 'launch'. So I added README.md & removed the keyboard_listener package.